Introduce location, product and filter routeParams
It is now possible to share a complete link with the
current location and type/product selection.

The '@' parameter specifies the geolocation in the
format: zoom,latitude,longitude

'products' is a CSV string with product IDs and
'types' is a CSV string with possible dealer types.

If the selection is 'all', products and types is not
set (i.e. not visible in the URL)

ref: #5

Signed-off-by: Maximilian Güntner <maximilian.guentner@gmail.com>